Added support for project (repository) creation from templates.
The project templates are simply stored in a repository like any other
project. Each first level directory in the repository is a template,
which underlying hierarchy (files, subfolders and SVN properties) are
copied to the projects created from that template.

Because the templates are stored in a SVN repository, SVN can be defined
on the template's files and directories (e.g. svn::global-ignore,
tsvn:logminsize, tsvn:projectlanguage, etc.).

Files and folders that contain in the placeholder #ProjectName# in their
names are renamed by subsituting the placeholder with the project's
name.

Access control can be specified by putting a special file named
'svnaccess.json' in the template's top directory. This file (which is
not  copied to projects created from the template) has JSON format, and
specifies the project specific groups that shall be created and the
access rights to these groups (and also to pre-existing groups if
needed).

Besides that, unit tests have begun to be migrated to their own
hierarchy under src/tests directory, and some have been fixed to start
working again.